Update graphite lib to 1.3.7+
- Also clean up leftover dead files that weren't cleaned up by Mozilla's graphite updates
- Also revert BMO 1255731 which was just a cheap solution to not back-port graphite 1.3.7 to ESR38

+++ gfx/graphite2/src/inc/Face.h
@@ -26,7 +26,7 @@ of the License or (at your option) any later version.
 */
 #pragma once
 
-#include <cstdio>
+#include <stdio.h>
 
 #include "graphite2/Font.h"
 
@@ -87,7 +87,7 @@ public:
     const FeatureRef  * feature(uint16 index) const;
 
     // Glyph related
-    uint16 getGlyphMetric(uint16 gid, uint8 metric) const;
+    int32  getGlyphMetric(uint16 gid, uint8 metric) const;
     uint16 findPseudo(uint32 uid) const;
 
     // Errors
